requirements.txt
feedparser
newspaper3k==0.2.8
lxml_html_clean==0.4.2
rank-bm25==0.2.2
PyYAML>=6.0
transformers>=4.41,<5
torch>=2.2,<3
sentencepiece>=0.1.99
protobuf>=3.20,<6
trafilatura>=1.8
accelerate>=0.30
